backend/requirements.txt
# Core FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and caching
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# AI and LangChain (compatible versions)
openai==1.3.5
langchain==0.1.17
langchain-core==0.1.52
langchain-openai==0.1.8
langsmith==0.1.17

# Authentication and security
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Updated Pydantic for compatibility
pydantic==2.7.4
pydantic-settings==2.3.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3

# Background tasks
celery==5.3.4

# Testing
pytest==7.4.3
httpx==0.25.2

# Additional utilities
python-dateutil==2.8.2
typing-extensions==4.8.0